Add input validation for player's choice
Player can (case insensitive) enter any of the three choices. If the
player enters something else, they are given four more tries before the
game gives up. After the first try, subsequent prompts also tell the
player how to exit the game (typing 'exit' into the prompt) so that the
game is not accidentally taking over the player's browser. If, for some
reason, the player isn't able to enter even that without typos, after
the fifth attempt the game ends.