Refactor playRound and create evaluateRound
The playRound function now manages the entire game round logic. All it
needs is the player's choice. Putting all of this logic into a function
makes the event function for the buttons much simpler.

Because the name 'playRound' was already taken for the function that
carried out the rock-paper-scissors logic, I had to rename that function
to evaluateRound.

We now have that playRound generates the computer's guess, calls
evaluateRound, and then finally calls reportRound. Round logic is no
longer scattered around different events.